requirements.txt
streamlit==1.29.0
python-dotenv==1.0.0
Pillow==10.1.0
pandas==2.1.4
openpyxl==3.1.2
minio==7.2.0
requests==2.31.0
